requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sentence-transformers
groq
qdrant-client
